# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/v1/years/": {
            "post": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Create a PersianYear",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "PersianYears"
                ],
                "summary": "Create a PersianYear",
                "parameters": [
                    {
                        "description": "Create a PersianYear",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_dto.CreatePersianYearRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "PersianYear response",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_dto.PersianYearResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse"
                        }
                    }
                }
            }
        },
        "/v1/years/get-by-filter": {
            "post": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Get PersianYears",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "PersianYears"
                ],
                "summary": "Get PersianYears",
                "parameters": [
                    {
                        "description": "Request",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_domain_filter.PaginationInputWithFilter"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PersianYear response",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_domain_filter.PagedList-github_com_naeemaei_golang-clean-web-api_api_dto_PersianYearResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse"
                        }
                    }
                }
            }
        },
        "/v1/years/{id}": {
            "get": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Get a PersianYear",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "PersianYears"
                ],
                "summary": "Get a PersianYear",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PersianYear response",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_dto.PersianYearResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Update a PersianYear",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "PersianYears"
                ],
                "summary": "Update a PersianYear",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update a PersianYear",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_dto.UpdatePersianYearRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PersianYear response",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_dto.PersianYearResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Delete a PersianYear",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "PersianYears"
                ],
                "summary": "Delete a PersianYear",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "response",
                        "schema": {
                            "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_naeemaei_golang-clean-web-api_api_dto.CreatePersianYearRequest": {
            "type": "object",
            "properties": {
                "endAt": {
                    "type": "string"
                },
                "persianTitle": {
                    "type": "string",
                    "maxLength": 4,
                    "minLength": 4
                },
                "startAt": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_naeemaei_golang-clean-web-api_api_dto.PersianYearResponse": {
            "type": "object",
            "properties": {
                "endAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "persianTitle": {
                    "type": "string"
                },
                "startAt": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_naeemaei_golang-clean-web-api_api_dto.UpdatePersianYearRequest": {
            "type": "object",
            "properties": {
                "endAt": {
                    "type": "string"
                },
                "persianTitle": {
                    "type": "string",
                    "maxLength": 4,
                    "minLength": 4
                },
                "startAt": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse": {
            "type": "object",
            "properties": {
                "error": {},
                "result": {},
                "resultCode": {
                    "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.ResultCode"
                },
                "success": {
                    "type": "boolean"
                },
                "validationErrors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_validation.ValidationError"
                    }
                }
            }
        },
        "github_com_naeemaei_golang-clean-web-api_api_helper.ResultCode": {
            "type": "integer",
            "enum": [
                0,
                40001,
                40101,
                40301,
                40401,
                42901,
                42902,
                50001,
                50002
            ],
            "x-enum-varnames": [
                "Success",
                "ValidationError",
                "AuthError",
                "ForbiddenError",
                "NotFoundError",
                "LimiterError",
                "OtpLimiterError",
                "CustomRecovery",
                "InternalError"
            ]
        },
        "github_com_naeemaei_golang-clean-web-api_api_validation.ValidationError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "property": {
                    "type": "string"
                },
                "tag": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "github_com_naeemaei_golang-clean-web-api_domain_filter.Filter": {
            "type": "object",
            "properties": {
                "filterType": {
                    "description": "text number",
                    "type": "string"
                },
                "from": {
                    "type": "string"
                },
                "to": {
                    "type": "string"
                },
                "type": {
                    "description": "contains notContains equals notEqual startsWith lessThan lessThanOrEqual greaterThan greaterThanOrEqual inRange endsWith",
                    "type": "string"
                }
            }
        },
...
        "github_com_naeemaei_golang-clean-web-api_domain_filter.PagedList-github_com_naeemaei_golang-clean-web-api_api_dto_PersianYearResponse": {
            "type": "object",
            "properties": {
                "hasNextPage": {
                    "type": "boolean"
                },
                "hasPreviousPage": {
                    "type": "boolean"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_dto.PersianYearResponse"
                    }
                },
                "pageNumber": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "totalPages": {
                    "type": "integer"
                },
                "totalRows": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_naeemaei_golang-clean-web-api_domain_filter.PaginationInputWithFilter": {
            "type": "object",
            "properties": {
                "filter": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_domain_filter.Filter"
                    }
                },
                "pageNumber": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "sort": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_domain_filter.Sort"
                    }
                }
            }
        },
        "github_com_naeemaei_golang-clean-web-api_domain_filter.Sort": {
            "type": "object",
            "properties": {
                "colId": {
                    "type": "string"
                },
                "sort": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "AuthBearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}